refactor: extract the circuit viewer into its own feature module
The 3D circuit viewer lives under `scan-config/components/`, but it is
not scan-config specific — the synaptome (beta) detail page needs the
same component, and a detail page has no business importing out of a
config-editor feature.

Moves it to `src/features/circuit-viewer/`:

  circuit-preview.tsx, circuit-viz/, color-by/, large-circuit-preview/,
  shared/3d-viewer.ts, shared/visualization-loading-indicator.tsx,
  hooks/circuit.ts -> hooks/use-circuit-image-url.ts

The viewer-owned zod schemas (Node/Nodes/Section/Sections, Cell,
MorphoViewerTreeItem) move out of `scan-config/types.ts` into
`circuit-viewer/types.ts`. Without that the new module would import back
up into scan-config and the extraction would be cosmetic.

Adds `'use client'` to `circuit-preview.tsx` and `circuit-viz.tsx`.
Today they get away without it because their only consumer pulls them
through `dynamic(..., { ssr: false })`; the detail page renders from an
async Server Component, so the boundary has to be declared on the module
itself.

Also drops `circuit-viz/debounced-switch/`, which has no importers.

No behaviour change. Vitest 437 passed, `next build` succeeds, no new
`tsc` errors (222 vs 227 pre-existing on main), barrel violations
2540 -> 2538.
