+++ src/features/circuit-viewer/color-by/palette.ts
