+++ src/features/circuit-viewer/color-by/properties.ts
