+++ src/features/circuit-viewer/color-by/use-fullscreen-element.ts
