+++ src/features/circuit-viewer/color-by/use-node-color-mapping.ts
