+++ src/features/circuit-viewer/hooks/use-circuit-image-url.ts
