+++ src/features/circuit-viewer/index.ts
@@ -1,0 +1,3 @@
+export { CircuitPreview } from './circuit-preview';
+
+export type { ViewerTheme } from './color-by/contrast';
