+++ src/features/circuit-viewer/large-circuit-preview/hooks.ts
