+++ src/features/circuit-viewer/large-circuit-preview/large-circuit-preview.module.css
