+++ src/features/circuit-viewer/shared/visualization-loading-indicator.tsx
