--- src/features/scan-config/components/circuit-viz/debounced-switch/index.ts
@@ -1,1 +1,0 @@
-export { default } from './debounced-switch';
